KIT-460: Update function & selector name, break file into smaller pieces
https://coveord.atlassian.net/browse/KIT-460
