+++ packages/atomic/cypress.json
@@ -1,6 +1,6 @@
 {
   "integrationFolder": "cypress/integration",
-  "ignoreTestFiles": "*selectors.ts",
+  "ignoreTestFiles": ["*selectors.ts", "*utils.ts"],
   "experimentalFetchPolyfill": true,
   "reporter": "spec",
   "viewportHeight": 1080,
